Check asset restrictions - added check asset restrictions while pay reward - added unit test

+++ tests/common/database_fixture.hpp
@@ -221,7 +221,7 @@ struct database_fixture {
       const std::string& name,
       const account_object& registrar,
       const account_object& referrer,
-      uint8_t referrer_percent = 100,
+      uint16_t referrer_percent = 100,
       public_key_type key = public_key_type()
       );
 
@@ -300,7 +300,7 @@ struct database_fixture {
       const string& name,
       const account_object& registrar,
       const account_object& referrer,
-      uint8_t referrer_percent = 100,
+      uint16_t referrer_percent = 100,
       const public_key_type& key = public_key_type()
       );
 
@@ -309,7 +309,7 @@ struct database_fixture {
       const private_key_type& key,
       const account_id_type& registrar_id = account_id_type(),
       const account_id_type& referrer_id = account_id_type(),
-      uint8_t referrer_percent = 100
+      uint16_t referrer_percent = 100
       );
 
    const committee_member_object& create_committee_member( const account_object& owner );
